Fix AC_CONFIG_MACRO_DIR
libtoolize parses configure.ac with sed directly.
This commit causes it to process it properly with
m4 first to include the includes that we do.
This causes AC_CONFIG_MACRO_DIR to start working.

Set an appropriate AUTOM4TE environment variable
(this is used by libtoolize as well).

Move the generated m4 files to m4/.